chamfer_distance added

+++ train-mmnist.py
@@ -83,8 +83,8 @@ def train():
         data = data.to(device)
         data = augmentation_transformer(data)
         optimizer.zero_grad()
-        out = model(data)
-        labels = data.y[:, 1:]
+        out = model(data).float()
+        labels = data.y[:, 1:].float()
         pc_shape = (data.num_graphs, data.num_nodes // data.num_graphs, 3)
         dist1, dist2 = criterion(out.reshape(pc_shape), labels.reshape(pc_shape))
         loss = (torch.mean(dist1)) + (torch.mean(dist2))
